Experiment with whole surface SiPM for crystal

+++ src/Utils.cc
@@ -381,7 +381,7 @@ void Utils::ApplyScintillation(G4Material* mat,
         const auto it = c.find("SCINTILLATIONYIELD");
         if (it == c.end())
             throw std::runtime_error("Missing SCINTILLATIONYIELD in const map");
-        mpt->AddConstProperty("SCINTILLATIONYIELD", it->second);
+        mpt->AddConstProperty("SCINTILLATIONYIELD", static_cast<int>(it->second));
     }
 
     AddConstIfPresent(mpt, c, "RESOLUTIONSCALE");
